Fix OffsetT type to be `long long` to fix Windows MSVC.
`long` on Linux is 64-bit, but 32-bit on Windows.  `long long` is 64-bit
on both.

+++ c/parallel/src/transform.cu
@@ -38,7 +38,7 @@
 struct op_wrapper;
 struct device_transform_policy;
 
-using OffsetT = long;
+using OffsetT = long long;
 static_assert(std::is_same_v<cub::detail::choose_signed_offset_t<OffsetT>, OffsetT>,
               "OffsetT must be signed int32 or int64");
